requirements_cc.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
arrow==1.4.0
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
blessed==1.22.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cmake==3.31.0
contourpy
croniter==1.4.1
cycler==0.12.1
dateutils==0.6.12
deepdiff==7.0.1
deepspeed
distro==1.9.0
docstring_parser==0.17.0
editor==1.7.0
einops==0.8.0
exceptiongroup==1.3.1
fastapi
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.46
glcontext==3.0.0
glfw==1.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio
grpcio-status
gym==0.26.2
gym-notices==0.1.0
gym3==0.3.3
h11==0.16.0
hf-xet==1.4.2
hjson==3.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg
importlib_metadata==8.7.1
importlib_resources==6.5.2
inquirer==3.4.1
internetarchive==5.8.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonargparse==4.27.7
jsonnet==0.21.0
jsonpatch==1.33
jsonpointer==3.0.0
kaleido==0.2.1
kiwisolver==1.5.0
kmeans-pytorch==0.3
kornia==0.7.4
lightning==2.0.8
lightning-cloud==0.6.0
lightning-utilities==0.15.3
lit==18.1.8
llvmlite==0.46.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib
mdurl==0.1.2
moderngl==5.12.0
mpmath==1.3.0
multidict==6.7.1
natsort==8.4.0
networkx==3.4.2
ninja==1.13.0
numba
numpy==1.26.4
omegaconf==2.3.0
open_clip_torch==3.3.0
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==24.2
pandas
pillow==12.1.1
piq==0.8.0
platformdirs==4.9.4
plotly==5.24.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==6.1.1
ptlflow==0.4.1
py-cpuinfo==9.0.0
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic
pydantic_core
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytorch-lightning==2.0.1
pytz==2026.1.post1
PyYAML==6.0.3
readchar==4.2.1
regex==2026.2.28
requests==2.32.5
rich==14.3.3
runs==1.3.0
safetensors
scikit-learn
scipy==1.14.1
sentry-sdk==2.55.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.50.0
starsessions==1.3.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
timm==1.0.25
tokenizers
torch==2.1.1
torchmetrics
torchvision
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
triton
typer==0.24.1
typeshed_client==2.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wandb
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
xformers
xmod==1.9.0
yarl==1.23.0
youtube-dl==2021.12.17
yt-dlp==2026.3.13
zipp==3.23.0
